+++ Cesium3DTilesSelection/test/MockTilesetContentManager.cpp
@@ -3,6 +3,7 @@
 #include <Cesium3DTilesSelection/Tile.h>
 
 namespace Cesium3DTilesSelection {
+
 void MockTilesetContentManagerTestFixture::setTileLoadState(
     Cesium3DTilesSelection::Tile& tile,
     Cesium3DTilesSelection::TileLoadState loadState) {
@@ -14,4 +15,11 @@ void MockTilesetContentManagerTestFixture::setTileShouldContinueUpdating(
     bool shouldContinueUpdating) {
   tile.setMightHaveLatentChildren(shouldContinueUpdating);
 }
+
+void MockTilesetContentManagerTestFixture::setTileContent(
+    Cesium3DTilesSelection::Tile& tile,
+    Cesium3DTilesSelection::TileContent&& content) {
+  tile._content = std::move(content);
+};
+
 } // namespace Cesium3DTilesSelection